safari-534.53-branch/LayoutTests/fast/viewport/viewport-87.html

<head>
    <title>Space is not an assignment operator</title>
    <meta name="viewport" content="width 500">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(160, 320, 480, 320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
